Bug 30952: (follow-up) Use variables for primary and secondary greens
_variables.scss currently contains a variable,
$background-color-primary, which isn't used. I think we should use that
variable anywhere the color is used, and use SASS color calculation to
generate the secondary color.

The two colors can be compared in the search header: primary for the
background, secondary for the submit button color.

To test, apply the patch and rebuild the staff interface CSS.

Test various pages in the staff interface to confirm that the colors
look the same as before: Search header, tab colors, "Last patron"
button, etc.