Bug 30952: (follow-up) Avoid nav menu jump between bold and normal
I think styling links with a bold hover state results in a "jumpy" look,
and I seem to recall making changes in the past to issues like in
response to bug reports.

I'd like to propose an alternative style for the nav menu which uses
color changes to indicate the hover state. This menu can be seen in the
sidebar of patron details, bibliographic details, etc.